reload

+++ data.py
@@ -17,11 +17,10 @@ def save_paths(src_paths, f_path):
 def cls2onehot(cls, depth):
     if not type(cls).__module__ == np.__name__:
         cls=np.asarray(cls)
-    cls=cls.astype(int)
 
-
+    cls=cls.astype(np.int32)
     debug_flag = False
-    labels = np.zeros([len(cls), 2])
+    labels = np.zeros([len(cls), 2] , dtype=np.int32)
     for i, ind in enumerate(cls):
         labels[i][ind:ind + 1] = 1
     if __debug__ == debug_flag:
